begining of the end

+++ app/Http/Controllers/UserController.php
@@ -9,6 +9,7 @@ use Illuminate\Validation\Rule;
 use Illuminate\Support\Facades\Storage;
 use App\Models\Idea;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use App\Models\Game;
 
 class UserController extends Controller
 {
@@ -21,7 +22,9 @@ class UserController extends Controller
     {
         $ideas = $user->ideas()->paginate(5);
 
-        return view('users.show', compact('user', 'ideas'));
+        $games = Game::all();
+
+        return view('users.show', compact('user', 'ideas', 'games'));
     }
 
 
@@ -34,8 +37,9 @@ class UserController extends Controller
 
         $editing = true;
         $ideas = $user->ideas()->paginate(5);
+        $games = Game::all();
 
-        return view('users.edit', compact('user', 'editing', 'ideas'));
+        return view('users.edit', compact('user', 'editing', 'ideas', 'games'));
 
     }
 
@@ -43,24 +47,28 @@ class UserController extends Controller
      * Update the specified resource in storage.
      */
     public function update(UpdateUserRequest $request, User $user)
-{
-    $validated = $request->validated();
-
-    if($request->hasFile('image')) // Check if a new image is uploaded
     {
-        // Delete the old image only if a new one is uploaded
-        Storage::disk('public')->delete($user->image);
+        $validated = $request->validated();
 
-        // Store the new image and update the validated array with the new path
-        $imagePath = $request->file('image')->store('profile', 'public');
-        $validated['image'] = $imagePath;
+        if ($request->hasFile('image')) // Check if a new image is uploaded
+        {
+            // Check if the user already has an image and it is not null
+            if ($user->image) {
+                // Delete the old image only if it exists
+                Storage::disk('public')->delete($user->image);
+            }
+
+            // Store the new image and update the validated array with the new path
+            $imagePath = $request->file('image')->store('profile', 'public');
+            $validated['image'] = $imagePath;
+        }
+
+        // Update the user's information with the validated data
+        $user->update($validated);
+
+        return redirect()->route('users.show', $user);
     }
 
-    // Update the user's information with the validated data
-    $user->update($validated);
-
-    return redirect()->route('users.show', $user);
-}
 
 
 }
